[refactoring/coupong-21] Chat 관련 4-Layered Architecture 적용
