+++ src/main/java/com/onepage/coupong/business/chat/dto/FilteringRequestDto.java
@@ -1,0 +1,8 @@
+package com.onepage.coupong.business.chat.dto;
+
+import lombok.Getter;
+
+@Getter
+public class FilteringRequestDto {
+    private String message;
+}
