--- src/main/java/com/onepage/coupong/chat/dto/FilteringRequestDTO.java
@@ -1,8 +1,0 @@
-package com.onepage.coupong.chat.dto;
-
-import lombok.Getter;
-
-@Getter
-public class FilteringRequestDTO {
-    private String message;
-}
